feat: add cacheHydrate util and apply to resolvers
